clean makefile clean rule

+++ Makefile
@@ -6,6 +6,7 @@ CC		=	gcc
 CFLAGS	=	-Wall -Wextra -Werror -O3 -g
 RM		=	rm -rf
 TESTER	=	./rsc/nm_tester.sh 
+CALL_TESTER = ./rsc/call_tester.sh
 DISPLAY_ASCII_ART = ./rsc/mk/ascii.sh
 
 ifeq ($(findstring bonus, $(MAKECMDGOALS)), bonus)
@@ -35,30 +36,31 @@ $(OBJ_DIR)/%.o: $(SRC_DIR)/%.c
 	@[ -d $(OBJ_DIR) ] || mkdir -p $(OBJ_DIR)
 	@$(CC) -o $@ -c $< $(CFLAGS)
 
-
-# TEST RULE
 test :	$(NAME)
 	@./$(NAME) $(NAME)
 
 dtest:	$(NAME)
 	@$(TESTER) $(NAME) 0
 
-all_test:
-	@./rsc/call_tester.sh
+all_test:	bonus
+	@$(CALL_TESTER)
 
 bonus:	$(NAME)
 
 clean:
 ifeq ($(shell [ -d $(OBJ_DIR) ] && echo 0 || echo 1), 0)
-	@printf "$(RED)Clean $(OBJ_DIR)$(RESET)\n"
 	@$(RM) $(OBJ_DIR)
-	@printf "$(RED)Clean $(NAME)$(RESET)\n"
+	@printf "$(RED)Clean $(OBJ_DIR)/test output$(RESET)\n"
 endif
+	@$(RM) nm_out out valgrind_out
 
 fclean:		clean
+ifeq ($(shell [ -d $(NAME) ] && echo 0 || echo 1), 0)
+	@$(RM) $(NAME)
+	@printf "$(RED)Clean $(NAME)$(RESET)\n"
+endif
 	@$(MAKE_LIBFT) fclean
 	@$(MAKE_LIST) fclean
-	@$(RM) $(NAME)
 
 re:			fclean all
